[CPDLP-1150] - refactor
