+++ app/controllers/finance/npq/statements_controller.rb
@@ -14,10 +14,10 @@ module Finance
       end
 
       def voided
-        @npq_lead_provider = lead_provider_scope.find(params[:lead_provider_id])
-        @cpd_lead_provider = @npq_lead_provider.cpd_lead_provider
-        @statement         = @cpd_lead_provider.npq_lead_provider.statements.find_by(name: identifier_to_name)
-        @voided_declarations = @statement.participant_declarations.voided
+        @npq_lead_provider   = lead_provider_scope.find(params[:lead_provider_id])
+        @cpd_lead_provider   = @npq_lead_provider.cpd_lead_provider
+        @statement           = @cpd_lead_provider.npq_lead_provider.statements.find_by(name: identifier_to_name)
+        @voided_declarations = ParticipantDeclaration::NPQ.where(statement: @statement).voided
       end
 
     private
