--- app/models/concerns/finance/statement_declaration_accountability.rb
@@ -1,3 +1,0 @@
-module Finance::Statement::DeclarationAccountability
-
-end
