+++ app/views/finance/ecf/participant_declarations/voided/show.html.erb
