+++ app/views/finance/npq/participant_declarations/voided/show.html.erb
